+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/attributes/EnergyNetComponent.java
@@ -66,6 +66,12 @@ public interface EnergyNetComponent extends ItemAttribute {
      */
     default int getCharge(@Nonnull Location l) {
         Validate.notNull(l, "Location was null!");
+
+        // Emergency fallback, this cannot hold a charge, so we'll just return zero
+        if (!isChargeable()) {
+            return 0;
+        }
+
         String charge = BlockStorage.getLocationInfo(l, "energy-charge");
 
         if (charge != null) {