+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/networks/energy/EnergyNetComponentType.java
@@ -4,6 +4,7 @@ import org.bukkit.block.Block;
 
 import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.Capacitor;
+import io.github.thebusybiscuit.slimefun4.implementation.items.electric.EnergyConnector;
 import io.github.thebusybiscuit.slimefun4.implementation.items.electric.reactors.Reactor;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AContainer;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AGenerator;
@@ -13,6 +14,7 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.abstractItems.AGenerator;
  * can have.
  * 
  * @author TheBusyBiscuit
+ * @author Linox
  * 
  * @see EnergyNetComponent
  * @see EnergyNet
@@ -37,6 +39,12 @@ public enum EnergyNetComponentType {
      * or other types of machinery.
      */
     CONSUMER,
+
+    /**
+     * A Connector transmits energy through the network.
+     * Also see: {@link EnergyConnector}
+     */
+    CONNECTOR,
 
     /**
      * A fallback value to use when a {@link Block} cannot be classified as any of the