+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/EnergyConnector.java
@@ -1,4 +1,4 @@
-package io.github.thebusybiscuit.slimefun4.implementation.items.cargo;
+package io.github.thebusybiscuit.slimefun4.implementation.items.electric;
 
 import javax.annotation.Nonnull;
 
@@ -7,16 +7,28 @@ import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.cscorelib2.chat.ChatColors;
+import io.github.thebusybiscuit.slimefun4.core.attributes.EnergyNetComponent;
 import io.github.thebusybiscuit.slimefun4.core.handlers.BlockUseHandler;
-import io.github.thebusybiscuit.slimefun4.core.networks.cargo.CargoNet;
+import io.github.thebusybiscuit.slimefun4.core.networks.energy.EnergyNet;
+import io.github.thebusybiscuit.slimefun4.core.networks.energy.EnergyNetComponentType;
 import io.github.thebusybiscuit.slimefun4.implementation.items.SimpleSlimefunItem;
 import me.mrCookieSlime.Slimefun.Lists.RecipeType;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.api.SlimefunItemStack;
 
-public class CargoConnectorNode extends SimpleSlimefunItem<BlockUseHandler> {
+/**
+ * This {@link EnergyNetComponent} is a connector for the {@link EnergyNet} networks.
+ * They work similar to {@link Capacitor capacitors}.
+ *
+ * @author Linox
+ *
+ * @see EnergyNet
+ * @see EnergyNetComponent
+ *
+ */
+public class EnergyConnector extends SimpleSlimefunItem<BlockUseHandler> implements EnergyNetComponent {
 
-    public CargoConnectorNode(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe, ItemStack recipeOutput) {
+    public EnergyConnector(Category category, SlimefunItemStack item, RecipeType recipeType, ItemStack[] recipe, ItemStack recipeOutput) {
         super(category, item, recipeType, recipe, recipeOutput);
     }
 
@@ -31,11 +43,22 @@ public class CargoConnectorNode extends SimpleSlimefunItem<BlockUseHandler> {
             Player p = e.getPlayer();
             Block b = e.getClickedBlock().get();
 
-            if (CargoNet.getNetworkFromLocation(b.getLocation()) != null) {
+            if (EnergyNet.getNetworkFromLocation(b.getLocation()) != null) {
                 p.sendMessage(ChatColors.color("&7Connected: " + "&2\u2714"));
             } else {
                 p.sendMessage(ChatColors.color("&7Connected: " + "&4\u2718"));
             }
         };
+    }
+
+    @Nonnull
+    @Override
+    public EnergyNetComponentType getEnergyComponentType() {
+        return EnergyNetComponentType.CONNECTOR;
+    }
+
+    @Override
+    public int getCapacity() {
+        return 0;
     }
 }